infra/prototype/requirements.txt
# Reddit Legal Crawler Requirements

# Reddit API
praw>=7.7.0

# AWS Integration
boto3>=1.28.0
botocore>=1.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0